+++ openscenario/openscenario_interpreter/include/openscenario_interpreter/syntax/init.hpp
@@ -15,6 +15,7 @@
 #ifndef OPENSCENARIO_INTERPRETER__SYNTAX__INIT_HPP_
 #define OPENSCENARIO_INTERPRETER__SYNTAX__INIT_HPP_
 
+#include <nlohmann/json.hpp>
 #include <openscenario_interpreter/syntax/init_actions.hpp>
 
 namespace openscenario_interpreter
@@ -30,16 +31,30 @@ inline namespace syntax
  *  </xsd:complexType>
  *
  * -------------------------------------------------------------------------- */
-struct Init : public InitActions
+struct Init
 {
+  const InitActions actions;
+
   template <typename Node, typename Scope>
   explicit Init(const Node & node, Scope & scope)
-  : InitActions(readElement<InitActions>("Actions", node, scope))
+  : actions(readElement<InitActions>("Actions", node, scope))
   {
   }
 
-  using InitActions::endsImmediately;
+  template <typename... Ts>
+  decltype(auto) evaluate(Ts &&... xs)
+  {
+    return actions.evaluate(std::forward<decltype(xs)>(xs)...);
+  }
+
+  template <typename... Ts>
+  decltype(auto) endsImmediately(Ts &&... xs)
+  {
+    return actions.endsImmediately(std::forward<decltype(xs)>(xs)...);
+  }
 };
+
+nlohmann::json & operator<<(nlohmann::json &, const Init &);
 }  // namespace syntax
 }  // namespace openscenario_interpreter