+++ openscenario/openscenario_interpreter/include/openscenario_interpreter/syntax/init_actions.hpp
@@ -15,6 +15,7 @@
 #ifndef OPENSCENARIO_INTERPRETER__SYNTAX__INIT_ACTIONS_HPP_
 #define OPENSCENARIO_INTERPRETER__SYNTAX__INIT_ACTIONS_HPP_
 
+#include <nlohmann/json.hpp>
 #include <openscenario_interpreter/syntax/action.hpp>
 #include <openscenario_interpreter/syntax/private.hpp>
 #include <string>
@@ -37,22 +38,21 @@ inline namespace syntax
  * </xsd:complexType>
  *
  * -------------------------------------------------------------------------- */
-#define ELEMENT(TYPE) \
-  std::make_pair(#TYPE, [&](auto && node) { push_back(make<TYPE>(node, scope)); })
-
 struct InitActions : public Elements
 {
   template <typename Node, typename Scope>
   explicit InitActions(const Node & node, Scope & scope)
   {
     std::unordered_map<std::string, std::function<void(const Node & node)>> dispatcher{
-      ELEMENT(GlobalAction),
-      ELEMENT(UserDefinedAction),
-      ELEMENT(Private),
+      // clang-format off
+      std::make_pair("GlobalAction",      [&](auto && node) { return push_back(make<GlobalAction>     (node, scope)); }),
+      std::make_pair("UserDefinedAction", [&](auto && node) { return push_back(make<UserDefinedAction>(node, scope)); }),
+      std::make_pair("Private",           [&](auto && node) { return push_back(make<Private>          (node, scope)); })
+      // clang-format on
     };
 
     for (const auto & each : node.children()) {
-      const auto iter{dispatcher.find(each.name())};
+      const auto iter = dispatcher.find(each.name());
       if (iter != std::end(dispatcher)) {
         std::get<1> (*iter)(each);
       }
@@ -83,7 +83,7 @@ struct InitActions : public Elements
   }
 };
 
-#undef ELEMENT
+nlohmann::json & operator<<(nlohmann::json &, const InitActions &);
 }  // namespace syntax
 }  // namespace openscenario_interpreter