+++ openscenario/openscenario_interpreter/include/openscenario_interpreter/syntax/triggering_entities.hpp
@@ -15,6 +15,7 @@
 #ifndef OPENSCENARIO_INTERPRETER__SYNTAX__TRIGGERING_ENTITIES_HPP_
 #define OPENSCENARIO_INTERPRETER__SYNTAX__TRIGGERING_ENTITIES_HPP_
 
+#include <openscenario_interpreter/reader/element.hpp>
 #include <openscenario_interpreter/syntax/entity_ref.hpp>
 #include <openscenario_interpreter/syntax/triggering_entities_rule.hpp>
 #include <utility>
@@ -34,21 +35,25 @@ inline namespace syntax
  *  </xsd:complexType>
  *
  * -------------------------------------------------------------------------- */
-struct TriggeringEntities : public std::list<EntityRef>
+struct TriggeringEntities
 {
-  const TriggeringEntitiesRule quantify;
+  const TriggeringEntitiesRule triggering_entities_rule;
+
+  const std::list<EntityRef> entity_refs;
 
   template <typename Node, typename Scope>
   explicit TriggeringEntities(const Node & node, Scope & scope)
-  : std::list<EntityRef>(readElements<EntityRef, 1>("EntityRef", node, scope)),
-    quantify(readAttribute<TriggeringEntitiesRule>("triggeringEntitiesRule", node, scope))
+  : triggering_entities_rule(
+      readAttribute<TriggeringEntitiesRule>("triggeringEntitiesRule", node, scope)),
+    entity_refs(readElements<EntityRef, 1>("EntityRef", node, scope))
   {
   }
 
   template <typename... Ts>
   constexpr decltype(auto) operator()(Ts &&... xs) const
   {
-    return quantify(std::begin(*this), std::end(*this), std::forward<decltype(xs)>(xs)...);
+    return triggering_entities_rule.apply(
+      std::begin(entity_refs), std::end(entity_refs), std::forward<decltype(xs)>(xs)...);
   }
 };
 }  // namespace syntax